MNT #185 moved the class
