+++ hkl/diffract.py
@@ -5,8 +5,6 @@ diffract
 Common Support for diffractometers
 
 .. autosummary::
-
-    ~Constraint
     ~Diffractometer
 
 """
@@ -25,76 +23,11 @@ import pyRestTable
 
 from . import calc
 from . import __version__
+from .util import Constraint
 
 
-__all__ = """
-    Constraint
-    Diffractometer
-""".split()
+__all__ = ["Diffractometer", ]
 logger = logging.getLogger(__name__)
-
-
-class Constraint:
-    """
-    Limitations on acceptable positions from computed forward() solutions.
-
-    Parameters
-    ----------
-    low_limit : float
-        Minimum acceptable solution for position.
-    high_limit : float
-        Maximum acceptable solution for position.
-    value : float
-        Constant value when ``fit=False``.
-    fit : bool
-        ``True`` when axis will be fitted, otherwise, hold position to ``value``.
-
-
-    note: Patterned on collections.namedtuple
-    """
-
-    def __init__(self, low_limit, high_limit, value, fit):
-        self.low_limit = float(low_limit)
-        self.high_limit = float(high_limit)
-        self.value = float(value)
-        self.fit = bool(fit)
-
-        self._fields = "low_limit high_limit value fit".split()
-        # fmt: off
-        _fields = ", ".join(
-            name + "=" + repr(getattr(self, name))
-            for name in self._fields
-        )
-        # fmt: on
-        self._repr_fmt = f"({_fields})"
-
-    def _asdict(self):
-        "Return a new dict which maps field names to their values."
-        return dict(zip(self._fields, self))
-
-    class _ConstraintIterator:
-        "Iterator"
-
-        def __init__(self, constraint):
-            self.constraint = constraint
-            self._index = 0
-
-        def __next__(self):
-            if self._index < len(self.constraint._fields):
-                c = getattr(self.constraint, self.constraint._fields[self._index])
-                self._index += 1
-                return c
-            else:
-                raise StopIteration
-
-    def __iter__(self):
-        "Iterate through the fields."
-        return self._ConstraintIterator(self)
-
-    def __repr__(self):
-        "Return a nicely formatted representation string."
-        content = "(" + ", ".join([f"{k}={getattr(self, k)}" for k in self._fields]) + ")"
-        return self.__class__.__name__ + content
 
 
 class Diffractometer(PseudoPositioner):
